backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
passlib[bcrypt]          # This is the one causing the current error
python-jose[cryptography] # Also essential for JWTs
python-multipart
aiosqlite
pydantic-settings       # For your core/config.py